Move class to extractors namespace, fix build errors
