+++ TwitchLib.Client.Models/ChatMessage.cs
@@ -211,7 +211,7 @@ namespace TwitchLib.Client.Models
             }
 
             if (EmoteSet == null)
-                EmoteSet = new EmoteSet(null, Message);
+                EmoteSet = new EmoteSet(default(string), Message);
 
             // Check if display name was set, and if it wasn't, set it to username
             if (string.IsNullOrEmpty(DisplayName))
